+++ ui/main/game/galactic_war/cards/gwaio_start_paratrooper.js
@@ -14,7 +14,10 @@ define([
       return gwoCard.loadoutIcon(CARD.id);
     },
     describe: _.constant(
-      "!LOC:Contains no basic factories, just Lobs and Unit Cannons built by the commander. Halves the cost of both. All land units can be built from the Unit Cannon as they are unlocked."
+      loc(
+        "!LOC:Contains no basic factories, just Lobs and Unit Cannons built by the commander. Strike from the skies, brothers!"
+      ) +
+        " Halves the cost of both. All land units can be built from the Unit Cannon as they are unlocked."
     ),
     hint: function () {
       return {