+++ ui/main/game/galactic_war/cards/gwaio_upgrade_omega.js
@@ -5,8 +5,7 @@ define([
   return {
     visible: _.constant(true),
     describe: _.constant(
-      "!LOC:Omega Upgrade Tech replaces the battleship's underside laser with an SXX laser." +
-        "Doubles the rate of fire of the other weapons." // added separately to avoid breaking translations
+      "!LOC:Omega Upgrade Tech replaces the battleship's underside laser with an SXX laser. Doubles the rate of fire of the other weapons."
     ),
     summarize: _.constant("!LOC:Omega Upgrade Tech"),
     icon: _.constant(