+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - Outdated AI additions in tech cards
 - Hoarder Commander not being offered Advanced Defense Technology
+- Broken translations on loadout screen (some are a bit scuffed)
 
 ## v5.69.0 - 2024-08-08